requirements.txt
google-cloud-bigquery
google-cloud-bigquery-storage
polars
mysql-connector-python
sqlalchemy